requirements.txt
matplotlib==3.1.3
Keras==2.4.3
numpy==1.19.1
seaborn==0.10.1
streamlit==0.58.0
pandas==1.0.3
Pillow==7.2.0
scikit_learn==0.23.2
